refactor: removed layers that are incongruent with our goals
