+++ kinotic-persistence/src/main/java/org/kinotic/persistence/internal/PersistenceInitializer.java
@@ -37,13 +37,6 @@ public class PersistenceInitializer {
 
     @PostConstruct
     public void init(){
-        int numToDeploy = kinoticProperties.getMaxNumberOfCoresToUse();
-        log.info("{} Cores will be used for Persistence Endpoints", numToDeploy);
-        DeploymentOptions options = new DeploymentOptions().setInstances(numToDeploy);
-
-        vertx.deployVerticle(verticleFactory::createOpenApiVerticle, options);
-
-        vertx.deployVerticle(verticleFactory::createGqlVerticle, options);
 
         if(properties.isEnableStaticFileServer()){// only 1 web server verticle
             vertx.deployVerticle(verticleFactory::createWebServerNextVerticle, new DeploymentOptions());
