+++ kinotic-persistence/src/main/java/org/kinotic/persistence/internal/endpoints/PersistenceVerticleFactory.java
@@ -3,14 +3,8 @@ package org.kinotic.persistence.internal.endpoints;
 import io.vertx.ext.healthchecks.HealthChecks;
 import lombok.RequiredArgsConstructor;
 import org.kinotic.core.api.config.KinoticProperties;
-import org.kinotic.core.api.security.SecurityContext;
-import org.kinotic.core.api.security.SecurityService;
-import org.kinotic.persistence.api.config.PersistenceProperties;
 import org.kinotic.core.api.config.OidcSecurityServiceProperties;
-import org.kinotic.persistence.internal.endpoints.graphql.DelegatingGqlHandler;
-import org.kinotic.persistence.internal.endpoints.graphql.GqlVerticle;
-import org.kinotic.persistence.internal.endpoints.openapi.OpenApiVerticle;
-import org.kinotic.persistence.internal.endpoints.openapi.OpenApiVertxRouterFactory;
+import org.kinotic.persistence.api.config.PersistenceProperties;
 import org.springframework.stereotype.Component;
 import tools.jackson.databind.ObjectMapper;
 
@@ -24,29 +18,13 @@ public class PersistenceVerticleFactory {
 
     // Common Deps
     private final KinoticProperties kinoticProperties;
-    private final SecurityContext securityContext;
     private final PersistenceProperties properties;
-    private final SecurityService securityService;
-
-    // Open Api Deps
-    private final OpenApiVertxRouterFactory openApiVertxRouterFactory;
-
-    // Gql Deps
-    private final DelegatingGqlHandler delegatingGqlHandler;
 
     // Web Server Deps
     private final ObjectMapper objectMapper;
     private final HealthChecks healthChecks;
     private final OidcSecurityServiceProperties oidcSecurityServiceProperties;
 
-    
-    public GqlVerticle createGqlVerticle(){
-        return new GqlVerticle(delegatingGqlHandler, properties, kinoticProperties.getSsl(), securityService, securityContext);
-    }
-
-    public OpenApiVerticle createOpenApiVerticle(){
-        return new OpenApiVerticle(properties, kinoticProperties.getSsl(), openApiVertxRouterFactory.createRouter());
-    }
 
     public WebServerVerticle createWebServerNextVerticle(){
         return new WebServerVerticle(objectMapper, healthChecks, properties, kinoticProperties.getSsl(), oidcSecurityServiceProperties);
