--- kinotic-persistence/src/main/java/org/kinotic/persistence/internal/endpoints/graphql/DelegatingGqlHandler.java
@@ -1,13 +1,0 @@
-package org.kinotic.persistence.internal.endpoints.graphql;
-
-import io.vertx.core.Handler;
-import io.vertx.ext.web.RoutingContext;
-
-/**
- * Delegates to the correct Vertx {@link io.vertx.ext.web.handler.graphql.GraphQLHandler} based on the path of the route
- * Created by Navíd Mitchell 🤪 on 11/19/24.
- */
-public interface DelegatingGqlHandler extends Handler<RoutingContext> {
-
-
-}
