--- kinotic-persistence/src/main/java/org/kinotic/persistence/internal/endpoints/graphql/GqlCacheKey.java
@@ -1,6 +1,0 @@
-package org.kinotic.persistence.internal.endpoints.graphql;
-
-/**
- * Cache key for the GraphQL schema handler cache, keyed by organization and application.
- */
-public record GqlCacheKey(String organizationId, String applicationId) {}
