--- kinotic-persistence/src/main/java/org/kinotic/persistence/internal/endpoints/graphql/datafetchers/PagedQueryDataFetcher.java
@@ -1,88 +1,0 @@
-package org.kinotic.persistence.internal.endpoints.graphql.datafetchers;
-
-import org.kinotic.persistence.api.services.EntitiesRepository;
-import tools.jackson.databind.ObjectMapper;
-import graphql.ExecutionResult;
-import graphql.schema.DataFetcher;
-import graphql.schema.DataFetchingEnvironment;
-import io.vertx.ext.web.RoutingContext;
-import org.kinotic.core.api.crud.Page;
-import org.kinotic.core.api.crud.Pageable;
-import org.kinotic.idl.api.schema.FunctionDefinition;
-import org.kinotic.idl.api.schema.ParameterDefinition;
-import org.kinotic.persistence.api.model.EntityContext;
-import org.kinotic.persistence.api.model.EntityDefinition;
-import org.kinotic.persistence.api.model.idl.PageableC3Type;
-import org.kinotic.persistence.internal.api.services.sql.MapParameterHolder;
-import org.kinotic.persistence.internal.endpoints.openapi.RoutingContextToEntityContextAdapter;
-import org.kinotic.persistence.internal.utils.GqlUtils;
-
-import java.util.Map;
-import java.util.Objects;
-import java.util.concurrent.CompletableFuture;
-
-/**
- * Created by Navíd Mitchell 🤪 on 6/13/24.
- */
-@SuppressWarnings("rawtypes")
-public class PagedQueryDataFetcher implements DataFetcher<CompletableFuture<Page<Map>>> {
-
-    private final Pageable defaultPageable;
-    private final EntitiesRepository entitiesRepository;
-    private final ObjectMapper objectMapper;
-    private final String pageableName;
-    private final String queryName;
-    private final String entityDefinitionId;
-
-    /**
-     * Creates a {@link DataFetcher} that will execute a named query page and return the result as a {@link ExecutionResult}
-     * @param entitiesRepository the {@link EntitiesRepository} to use to execute the query
-     * @param objectMapper the {@link ObjectMapper} to use to deserialize the input parameters
-     * @param queryDefinition the {@link FunctionDefinition} for the query to execute
-     * @param defaultPageable the default {@link Pageable} to use if no pageable parameter is defined in the {@link FunctionDefinition}
-     * @param entityDefinitionId the id of the {@link EntityDefinition} that the query is for
-     */
-    public PagedQueryDataFetcher(EntitiesRepository entitiesRepository,
-                                 ObjectMapper objectMapper,
-                                 FunctionDefinition queryDefinition,
-                                 Pageable defaultPageable,
-                                 String entityDefinitionId) {
-        this.defaultPageable = defaultPageable;
-        this.entitiesRepository = entitiesRepository;
-        this.objectMapper = objectMapper;
-        this.queryName = queryDefinition.getName();
-        this.entityDefinitionId = entityDefinitionId;
-
-        String pageParamName = null;
-        for(ParameterDefinition parameter : queryDefinition.getParameters()){
-            if(parameter.getType() instanceof PageableC3Type){
-                pageParamName = parameter.getName();
-                break;
-            }
-        }
-        pageableName = pageParamName;
-    }
-
-    @Override
-    public CompletableFuture<Page<Map>> get(DataFetchingEnvironment environment) throws Exception {
-        RoutingContext rc = environment.getGraphQlContext().get(RoutingContext.class);
-        Objects.requireNonNull(rc);
-        EntityContext ec = new RoutingContextToEntityContextAdapter(rc);
-
-        Pageable pageable;
-        if(pageableName != null) {
-            pageable = GqlUtils.parseVariable(environment.getArguments(),
-                                              pageableName,
-                                              Pageable.class,
-                                              objectMapper);
-        }else {
-            pageable = defaultPageable;
-        }
-        return entitiesRepository.namedQueryPage(entityDefinitionId,
-                                                 queryName,
-                                                 new MapParameterHolder(environment.getArguments()),
-                                                 pageable,
-                                                 Map.class,
-                                                 ec);
-    }
-}
