--- kinotic-persistence/src/main/java/org/kinotic/persistence/internal/endpoints/openapi/OpenApiVerticle.java
@@ -1,45 +1,0 @@
-package org.kinotic.persistence.internal.endpoints.openapi;
-
-import io.vertx.core.Future;
-import io.vertx.core.VerticleBase;
-import io.vertx.core.http.HttpServer;
-import io.vertx.core.http.HttpServerOptions;
-import io.vertx.ext.web.Router;
-import lombok.RequiredArgsConstructor;
-import org.kinotic.core.api.config.SslHelper;
-import org.kinotic.core.api.config.SslProperties;
-import org.kinotic.persistence.api.config.PersistenceProperties;
-import org.kinotic.persistence.api.model.EntityDefinition;
-
-/**
- * We have one OpenApi spec for all {@link EntityDefinition}'s in an application. But this handles all applications and all EntityDefinitions.
- * Created by Navíd Mitchell 🤪 on 5/29/23.
- */
-@RequiredArgsConstructor
-public class OpenApiVerticle extends VerticleBase {
-
-    private final PersistenceProperties properties;
-    private final SslProperties sslProperties;
-    private final Router router;
-    private HttpServer server;
-
-
-    @Override
-    public Future<?> start() throws Exception {
-        HttpServerOptions options = new HttpServerOptions();
-        options.setMaxHeaderSize(properties.getMaxHttpHeaderSize());
-        SslHelper.applySsl(options, sslProperties);
-        server = vertx.createHttpServer(options);
-
-        // Begin listening for requests
-        return server.requestHandler(router)
-                     .listen(properties.getOpenApiPort());
-    }
-
-    @Override
-    public Future<?> stop() throws Exception {
-        return server.close();
-    }
-
-
-}
